Factor out DriverTraced into new module
