+++ marlowe-chain-sync/libchainsync/Language/Marlowe/Runtime/ChainSync.hs
@@ -24,8 +24,8 @@ import Language.Marlowe.Runtime.ChainSync.JobServer (ChainSyncJobServerDependenc
 import Language.Marlowe.Runtime.ChainSync.QueryServer (ChainSyncQueryServerDependencies(..), chainSyncQueryServer)
 import Language.Marlowe.Runtime.ChainSync.Server (ChainSyncServerDependencies(..), chainSyncServer)
 import Network.Protocol.Connection (SomeConnectionSourceTraced)
+import Network.Protocol.Driver.Trace (HasSpanContext)
 import Network.Protocol.Job.Server (JobServer)
-import Network.Protocol.Peer.Trace (HasSpanContext)
 import Network.Protocol.Query.Server (QueryServer)
 import Ouroboros.Network.Protocol.LocalTxSubmission.Client (SubmitResult)
 import UnliftIO (MonadUnliftIO)
