+++ marlowe-chain-sync/marlowe-chain-sync/Main.hs
@@ -42,7 +42,8 @@ import Language.Marlowe.Runtime.ChainSync.NodeClient (NodeClient(..), NodeClient
 import Logging (RootSelector(..), renderRootSelectorOTel)
 import Network.Protocol.ChainSeek.Server (chainSeekServerPeer)
 import Network.Protocol.Connection (SomeConnectionSourceTraced(..))
-import Network.Protocol.Driver (TcpServerDependencies(..), tcpServerTraced)
+import Network.Protocol.Driver (TcpServerDependencies(..))
+import Network.Protocol.Driver.Trace (tcpServerTraced)
 import Network.Protocol.Handshake.Server (handshakeConnectionSourceTraced)
 import Network.Protocol.Job.Server (jobServerPeer)
 import Network.Protocol.Query.Server (queryServerPeer)
