+++ marlowe-runtime/marlowe-indexer/Logging.hs
@@ -19,7 +19,7 @@ import Language.Marlowe.Runtime.Indexer (MarloweIndexerSelector(..))
 import Language.Marlowe.Runtime.Indexer.ChainSeekClient (ChainSeekClientSelector(..))
 import Language.Marlowe.Runtime.Indexer.Database.PostgreSQL (QueryField(..), QuerySelector(..))
 import Language.Marlowe.Runtime.Indexer.Store (ChangesStatistics(..), SaveField(..), StoreSelector(..))
-import Network.Protocol.Driver (TcpClientSelector, renderTcpClientSelectorOTel)
+import Network.Protocol.Driver.Trace (TcpClientSelector, renderTcpClientSelectorOTel)
 import Network.Protocol.Handshake.Types (Handshake)
 import Network.Protocol.Query.Types (Query)
 import Observe.Event (idInjectSelector, injectSelector)
