+++ marlowe-runtime/marlowe-tx/Main.hs
@@ -28,7 +28,8 @@ import qualified Language.Marlowe.Runtime.Transaction.Submit as Submit
 import Logging (RootSelector(..), renderRootSelectorOTel)
 import Network.Protocol.ChainSeek.Client (chainSeekClientPeer)
 import Network.Protocol.Connection (SomeClientConnectorTraced, SomeConnectionSourceTraced(..), SomeConnectorTraced(..))
-import Network.Protocol.Driver (TcpServerDependencies(..), runSomeConnectorTraced, tcpClientTraced, tcpServerTraced)
+import Network.Protocol.Driver (TcpServerDependencies(..))
+import Network.Protocol.Driver.Trace (runSomeConnectorTraced, tcpClientTraced, tcpServerTraced)
 import Network.Protocol.Handshake.Client (handshakeClientConnectorTraced)
 import Network.Protocol.Handshake.Server (handshakeConnectionSourceTraced)
 import Network.Protocol.Job.Client (jobClientPeer)
