+++ marlowe-runtime/sync/Language/Marlowe/Runtime/Sync/MarloweSyncServer.hs
@@ -15,8 +15,7 @@ import Language.Marlowe.Runtime.Core.Api (ContractId, MarloweVersion, SomeMarlow
 import Language.Marlowe.Runtime.History.Api (SomeCreateStep(..))
 import Language.Marlowe.Runtime.Sync.Database (DatabaseQueries(..), Next(..))
 import Network.Protocol.Connection (SomeConnectionSourceTraced, SomeServerConnectorTraced, acceptSomeConnectorTraced)
-import Network.Protocol.Driver (runSomeConnectorTraced)
-import Network.Protocol.Peer.Trace (HasSpanContext)
+import Network.Protocol.Driver.Trace (HasSpanContext, runSomeConnectorTraced)
 import UnliftIO (MonadUnliftIO)
 
 data MarloweSyncServerDependencies r s m = MarloweSyncServerDependencies
